RingBuffer: use the full capacity of the array
Before, to achieve a capacity of 16 elements RingBuffer used an array of 17
elements under the hood. Now the same can be accomplished with an array of 16
elements.

This simplifies the bounds of RingBuffer but it's technically a breaking change.

This change adds an overhead of 1 cycle to the enqueue and dequeue operations.

This commit also documents that using capacities that are powers of two results
in better performance (modulo operations become shifts). This was not documented
before and due to the mismatch of capacity and array size a RB with capacity of
15 performed better than one with a capacity of 16.