Remove implementing Deref<Target = [T]> for HistoryBuf
There is already `as_slice()` which gives access to the underlying
backing array where the elements are not in the order of writing.

With automatic dereferencing, functions like `iter()` or `windows()` are
automatically available for `HistoryBuf` and are returning items in an
order which does not reflect the write order and will likely surprise
users which did not carefully read all the documentation on for
`HistoryBuf` like me.