mpmc: Document #583 and add corresponding loom test
This patch documents #583 adds failing loom tests to exercise #583
hoping that a future algorithm change can make them pass.

The tests are ran in CI but their results is ignored for now as they fail.

Ideally the `loom` tests will cover more usage of `mpmc::Queue` in the future
and also cover `spsc`.